Improved test coverage for the project (#68)
* Add a call to the super class in MenuAction

* Added a test for the camera subsystem

* Added a test for the details builder

* Added another test for the other details panel

* Fixed tests to not use RTTI

* Added a test for the party screen

* Ran clang format

* Fixed remaining tests

* Fixed SonarCloud issues

* Fixed build issue

+++ Plugins/GridBased2D/Source/GridBased2D/Private/Components/CharacterBillboardAnimationComponent.cpp
@@ -88,8 +88,9 @@ void UCharacterBillboardAnimationComponent::SetUpMaterialInstance() {
         return;
     }
     BillboardComponent->Elements.Empty();
-    BillboardComponent->AddElement(MaterialInstance, nullptr, false, SourceTexture->GetSizeY() / (4 * 2),
-                                   SourceTexture->GetSizeX() / (4 * 2), nullptr);
+    BillboardComponent->AddElement(MaterialInstance, nullptr, false,
+                                   static_cast<float>(SourceTexture->GetSizeY() / (4 * 2)),
+                                   static_cast<float>(SourceTexture->GetSizeX() / (4 * 2)), nullptr);
 }
 
 UMaterialBillboardComponent *UCharacterBillboardAnimationComponent::GetBillboardComponent() const {
